backend/requirements.txt
rich
fastapi-limiter
pydantic
termcolor
requests
uvicorn
markdown
markdown-it-py
PyYAML
beautifulsoup4
asyncio
lxml
python-dotenv
aiohttp
slowapi
# image_coverter
pillow
Pillow-AVIF-Plugin